+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { Component, NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginPageComponent } from './pages/login-page/login-page.component';
-import { HomePageComponent } from './pages/home-page/home-page.component';
-import { SignupPageComponent } from './pages/signup-page/signup-page.component';
+import { LoginPageComponent } from './components/login-page/login-page.component';
+import { HomePageComponent } from './components/home-page/home-page.component';
+import { SignupPageComponent } from './components/signup-page/signup-page.component';
 
 const routes: Routes = [
   {path:'login', component:LoginPageComponent},
