+++ src/app/app.module.ts
@@ -4,7 +4,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { LoginPageComponent } from './pages/login-page/login-page.component';
+import { LoginPageComponent } from './components/login-page/login-page.component';
 import {MatCardModule} from '@angular/material/card';
 import {MatListModule} from '@angular/material/list';
 import {MatFormFieldModule} from '@angular/material/form-field';
@@ -15,8 +15,8 @@ import {MatIconModule} from '@angular/material/icon';
 import {MatButtonModule} from '@angular/material/button';
 import {MatDividerModule} from '@angular/material/divider';
 import { HttpClientModule } from '@angular/common/http';
-import { HomePageComponent } from './pages/home-page/home-page.component';
-import { SignupPageComponent } from './pages/signup-page/signup-page.component';
+import { HomePageComponent } from './components/home-page/home-page.component';
+import { SignupPageComponent } from './components/signup-page/signup-page.component';
 
 
 @NgModule({
